change msg and improved matlab func

+++ matlab/custom/matlab_msg_gen/glnxa64/install/include/interfaces/interfaces/msg/detail/detected__struct.h
@@ -23,7 +23,7 @@ typedef struct interfaces__msg__Detected
   int32_t color;
   int32_t x;
   int32_t y;
-  float angle;
+  int32_t angle;
 } interfaces__msg__Detected;
 
 // Struct for a sequence of interfaces__msg__Detected.
